Reorganize SSTATE_DIR, implement fallbacks
Structure SSTATE_DIR such that non-target sstates are placed in a directory
specific to the host we're running on, and allow fallback to sstates from
compatible hosts.

As one example of usage, at Mentor we set SSTATE_MIRROR_HOSTS =
"RedHatEnterprise-5", and set up an adjustment hook that ensures that rhel
server, rhel client, and centos all return that identifier. That's the oldest
distro we support, and we know native/cross binaries built there will run on
all the other distros we support.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>